+++ release-notes.md
@@ -4,6 +4,7 @@
 * Make hash functions configurable via application.conf
 * One more byte added to address to identify the network
 * Rpc address become configurable via settings.json 
+* Check, that transactions included in block are not too old added
 
 
 **1.2.7**